fix: move navbar to component
